Moved servicers to servicers/protected

<tej.prakash@wavelabs.ai>
